Remove duplicate code, improve stability, names
Duplicate code defined z_ and x_reconstructed_ twice. Doesn't get in the
way of computation, but was confusing.

Computation graph adapted to take advantage of TF implementation of
cross-entropy. This is a little faster and much more numerically stable.
Computation graph became a little bit more complicated, though, and
support for arbitrary squashing functions had to be dropped (since the
CE function has to match it). This feature might be restrictedly
reintroduced in the future if there's any need for it.