Simplify computation graph (implicit entry point)
It is not necessary to create a separate placeholder for inputting
z in decode. TF allows one to override a node's value directly.

By also reusing the exit point (now called x_out), the twin branch
of the decode graph can be eliminated. This simplifies the resulting
graph, and also the python code a bit.

There is one side-effect, though: default normal sampling is not
supported directly anymore. Since this operation is inexpensive,
computing it through ordinary code is a viable alternative.